Testing passive

+++ code/FiberPy/FiberPy/package/modules/characterize/characterize_model.py
@@ -42,10 +42,9 @@ def characterize_model(json_analysis_file_string):
     # If there is a manipulations section, use that to create the
     # appropriate models
     if ("manipulations" in anal_struct['model']):
+        print(json_analysis_file_string)
         json_analysis_file_string = \
             generate_model_files(json_analysis_file_string)
-            
-    print(json_analysis_file_string)
    
     # Pull off the characterization tasks
     char_struct = anal_struct['characterization']
@@ -171,7 +170,6 @@ def generate_model_files(json_analysis_file_string):
                     adj_model[a['variable']][a['isotype']-1]['state'][a['state']-1]['extension'] = \
                         value
                 else:
-                    print(a)
                     # Transition parameters
                     y = np.asarray(adj_model[a['variable']][a['isotype']-1]['state'][a['state']-1] \
                               ['transition'][a['transition']-1]['rate_parameters'],
@@ -206,7 +204,7 @@ def generate_model_files(json_analysis_file_string):
                 # Check for NaN
                 if (np.isnan(value)):
                     adj_model[a['class']][a['variable']] = 'null'
-    
+        
         # Now generate the model file string
         model_file_string = 'model_%i.json' % (i+1)
         
@@ -250,13 +248,15 @@ def deduce_pCa_length_control_properties(json_analysis_file_string,
     # Potentially switch off simulations
     figures_only = False
     if ('figures_only' in pCa_struct):
-        if (pCa_struct['figures_only'] == "True"):
+        if (pCa_struct['figures_only'] == 'True'):
             figures_only = True
-
-    trace_figures_on = True
-    if ('trace_figures_on' in pCa_struct):
-        if (pCa_struct['trace_figures_on'] == 'False'):
-            trace_figures_on = False
+            
+    figures_off = False
+    if ('figures_off' in pCa_struct):
+        if (pCa_struct['figures_off'] == 'True'):
+            figures_off = True
+            
+    trace_figures_on = False;
 
     # Load the file
     with open(json_analysis_file_string, 'r') as f:
@@ -602,7 +602,7 @@ def deduce_pCa_length_control_properties(json_analysis_file_string,
     # Deduce the output dir
     output_dir = str(Path(sim_output_dir).parent)
     
-    # Create a dict for functino return values
+    # Create a dict for function return values
     func_output = dict()
 
     # pCa curves
@@ -688,7 +688,9 @@ def deduce_pCa_length_control_properties(json_analysis_file_string,
         json.dump(pCa_lc_batch, f, indent=4)
         
     # Now run the isometric batch
-    batch.run_batch(pCa_lc_batch_file, figures_only=figures_only)
+    batch.run_batch(pCa_lc_batch_file,
+                    figures_only = figures_only,
+                    figures_off = figures_off)
     
     # Add the batch to the output
     func_output['sim_batch'] = pCa_lc_batch    
@@ -704,13 +706,15 @@ def deduce_fv_properties(json_analysis_file_string,
     # Potentially switch off simulations
     figures_only = False
     if ('figures_only' in fv_struct):
-        if (fv_struct['figures_only'] == "True"):
+        if (fv_struct['figures_only'] == 'True'):
             figures_only = True
+             
+    figures_off = False
+    if ('figures_off' in fv_struct):
+        if (fv_struct['figures_off'] == 'True'):
+            figures_off = True
 
-    trace_figures_on = True
-    if ('trace_figures_on' in fv_struct):
-        if (fv_struct['trace_figures_on'] == 'False'):
-            trace_figures_on = False
+    trace_figures_on = False 
 
     # Load the file
     with open(json_analysis_file_string, 'r') as f:
@@ -1001,16 +1005,36 @@ def deduce_fv_properties(json_analysis_file_string,
             # Update dir_counter
             dir_counter = dir_counter + 1
             
+            # Get the sim options file
+            if (model_struct['relative_to'] == 'this_file'):
+                model_dir = Path(json_analysis_file_string).parent.absolute()
+            else:
+                model_dir = ''
+                
+            # Copy the base options file
+            orig_options_file = os.path.join(model_dir,
+                                             model_struct['options_file'])
+            
+            with open(orig_options_file, 'r') as f:
+                orig_options_data = json.load(f)
+            
+            # Adjust the options struct if we have randomized repeats
+            if ('randomized_repeats' in fv_struct):
+                rand_repeats = fv_struct['randomized_repeats']
+                orig_options_data['options']['rand_seed'] = "random"
+            else:
+                rand_repeats = 1
+            
             if (fv_struct['pCa'] == "pCa_50"):
                 # Deduce data from the curve fit
                 c_fit = force_pCa_fit[dir_counter - 1]
                 isometric_force = c_fit['f_at_pCa_50']
                 isometric_job_index = (dir_counter - 1) * \
                                         len(fv_struct['pCa_values']) * \
-                                        fv_struct['randomized_repeats']
+                                        rand_repeats
                                         
             else:
-                # Pull off the isoric force for the preceding job
+                # Pull off the isometric force for the preceding job
                 isometric_job_index = dir_counter - 1
                 results_file_string = isometric_jobs[isometric_job_index]['results_file']
                 sim_data = pd.read_csv(results_file_string, sep='\t')
@@ -1052,13 +1076,6 @@ def deduce_fv_properties(json_analysis_file_string,
             orig_options_file = isometric_jobs[isometric_job_index]['options_file']
             with open(orig_options_file, 'r') as f:
                 orig_options_data = json.load(f)
-            
-            # Adjust the options struct if we have randomized repeats
-            if ('randomized_repeats' in fv_struct):
-                rand_repeats = fv_struct['randomized_repeats']
-                orig_options_data['options']['rand_seed'] = "random"
-            else:
-                rand_repeats = 1
     
             # Loop through the isotonic forces
             for (k, rel_f) in enumerate(fv_struct['rel_isotonic_forces']):
@@ -1247,7 +1264,9 @@ def deduce_fv_properties(json_analysis_file_string,
         json.dump(isotonic_batch, f, indent=4)
         
     # Now run the isotonic batch
-    batch.run_batch(isotonic_batch_file, figures_only=figures_only)
+    batch.run_batch(isotonic_batch_file,
+                    figures_only = figures_only,
+                    figures_off = figures_off)
 
 def deduce_freeform_properties(json_analysis_file_string,
                                freeform_struct):
@@ -1258,6 +1277,10 @@ def deduce_freeform_properties(json_analysis_file_string,
     if ('figures_only' in freeform_struct):
         if (freeform_struct['figures_only'] == "True"):
             figures_only = True
+
+    figures_off = False
+    if ('figures_off' in freeform_struct):
+        figures_off = freeform_struct['figures_off']    
 
     trace_figures_on = True
     if ('trace_figures_on' in freeform_struct):
@@ -1323,15 +1346,16 @@ def deduce_freeform_properties(json_analysis_file_string,
         prot_dir = os.path.join(base_dir,
                                 freeform_struct['protocol']['protocol_folder'])
         
-        # Clear it and then check it is there
-        try:
-            print('Trying to remove %s' % prot_dir)
-            shutil.rmtree(prot_dir, ignore_errors = True)
-        except OSError as e:
-            print('Error: %s : %s' % (prot_dir, e.strerror))
-            
-        if not os.path.isdir(prot_dir):
-            os.makedirs(prot_dir)
+        # If you are running simulations, delete the existing sim folders
+        if not figures_only:
+            try:
+                print('Trying to remove %s' % prot_dir)
+                shutil.rmtree(prot_dir, ignore_errors = True)
+            except OSError as e:
+                print('Error: %s : %s' % (prot_dir, e.strerror))
+                
+            if not os.path.isdir(prot_dir):
+                os.makedirs(prot_dir)
             
         # Set up for an array of protocols
         for (i, ps) in enumerate(freeform_struct['protocol']['data']):
@@ -1373,11 +1397,14 @@ def deduce_freeform_properties(json_analysis_file_string,
 
     # Clear the sim_folder and then check it is there
     sim_folder = os.path.join(base_dir, freeform_struct['sim_folder'])
-    try:
-        print('Trying to remove %s' % sim_folder)
-        shutil.rmtree(sim_folder, ignore_errors = True)
-    except OSError as e:
-        print('Error: %s : %s' % (sim_folder, e.strerror))
+    
+    # If you are running simulations, delete the existing sim folders
+    if not figures_only:
+        try:
+            print('Trying to remove %s' % sim_folder)
+            shutil.rmtree(sim_folder, ignore_errors = True)
+        except OSError as e:
+            print('Error: %s : %s' % (sim_folder, e.strerror))
         
     if not os.path.isdir(sim_folder):
         os.makedirs(sim_folder)
@@ -1527,9 +1554,6 @@ def deduce_freeform_properties(json_analysis_file_string,
                         freeform_prot_file = os.path.join(sim_input_dir,
                                                           fn)
                         
-                        print("opf: %s" % orig_prot_file)
-                        print("fpf: %s" % freeform_prot_file)
-                        
                         shutil.copyfile(orig_prot_file, freeform_prot_file)
     
                        # Create the job
@@ -1584,43 +1608,44 @@ def deduce_freeform_properties(json_analysis_file_string,
     # Now create the batch analysis section
     batch_figs = dict()
     
-    # Rates
-    batch_figs['rates'] = []
-    fig = dict()
-    fig['relative_to'] = "False"
-    fig['results_folder'] = os.path.join(base_dir,
-                                         freeform_struct['sim_folder'],
-                                         'sim_output')
-    fig['output_image_file'] = os.path.join(base_dir,
-                                            freeform_struct['sim_folder'],
-                                            'sim_output',
-                                            'rates')
-    fig['output_image_formats'] = freeform_struct['output_image_formats']
+    if (figures_off == False):
+        # Rates
+        batch_figs['rates'] = []
+        fig = dict()
+        fig['relative_to'] = "False"
+        fig['results_folder'] = os.path.join(base_dir,
+                                             freeform_struct['sim_folder'],
+                                             'sim_output')
+        fig['output_image_file'] = os.path.join(base_dir,
+                                                freeform_struct['sim_folder'],
+                                                'sim_output',
+                                                'rates')
+        fig['output_image_formats'] = freeform_struct['output_image_formats']
+        
+        if ('formatting' in freeform_struct):
+            fig['formatting'] = freeform_struct['formatting']
+        
+        batch_figs['rates'].append(fig)
     
-    if ('formatting' in freeform_struct):
-        fig['formatting'] = freeform_struct['formatting']
-    
-    batch_figs['rates'].append(fig)
-
-    # Superposed traces
-    batch_figs['superposed_traces'] = []
-    fig = dict()
-    fig['relative_to'] = "False"
-    fig['results_folder'] = os.path.join(base_dir,
-                                         freeform_struct['sim_folder'],
-                                         'sim_output')
-    fig['output_image_file'] = os.path.join(base_dir,
-                                            freeform_struct['sim_folder'],
-                                            'sim_output',
-                                            'superposed_traces')
-    fig['output_image_formats'] = freeform_struct['output_image_formats']
-    if ('superposed_x_ticks' in freeform_struct):
-        fig['superposed_x_ticks'] = freeform_struct['superposed_x_ticks']
-    if ('formatting' in freeform_struct):
-        fig['formatting'] = freeform_struct['formatting']        
-    batch_figs['superposed_traces'].append(fig)
-    
-    freeform_b['batch_figures'] = batch_figs
+        # Superposed traces
+        batch_figs['superposed_traces'] = []
+        fig = dict()
+        fig['relative_to'] = "False"
+        fig['results_folder'] = os.path.join(base_dir,
+                                             freeform_struct['sim_folder'],
+                                             'sim_output')
+        fig['output_image_file'] = os.path.join(base_dir,
+                                                freeform_struct['sim_folder'],
+                                                'sim_output',
+                                                'superposed_traces')
+        fig['output_image_formats'] = freeform_struct['output_image_formats']
+        if ('superposed_x_ticks' in freeform_struct):
+            fig['superposed_x_ticks'] = freeform_struct['superposed_x_ticks']
+        if ('formatting' in freeform_struct):
+            fig['formatting'] = freeform_struct['formatting']        
+        batch_figs['superposed_traces'].append(fig)
+        
+        freeform_b['batch_figures'] = batch_figs
                    
     # Now insert isometric_b into a full batch structure
     freeform_batch = dict()
